+++ darts/models/forecasting/arima.py
@@ -91,6 +91,8 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
     def _fit(self, series: TimeSeries, future_covariates: Optional[TimeSeries] = None):
         super()._fit(series, future_covariates)
 
+        self._assert_univariate(series)
+
         # storing to restore the statsmodels model results object
         self.training_historic_future_covariates = future_covariates
 
@@ -112,6 +114,7 @@ class ARIMA(TransferableFutureCovariatesLocalForecastingModel):
         historic_future_covariates: Optional[TimeSeries] = None,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ) -> TimeSeries:
 
         if num_samples > 1 and self.trend: