+++ darts/models/forecasting/croston.py
@@ -103,7 +103,7 @@ class Croston(FutureCovariatesLocalForecastingModel):
 
     def _fit(self, series: TimeSeries, future_covariates: Optional[TimeSeries] = None):
         super()._fit(series, future_covariates)
-        series._assert_univariate()
+        self._assert_univariate(series)
         series = self.training_series
 
         self.model.fit(
@@ -120,6 +120,7 @@ class Croston(FutureCovariatesLocalForecastingModel):
         n: int,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ):
         super()._predict(n, future_covariates, num_samples)
         values = self.model.predict(