+++ darts/models/forecasting/fft.py
@@ -272,6 +272,7 @@ class FFT(LocalForecastingModel):
     def fit(self, series: TimeSeries):
         series = fill_missing_values(series)
         super().fit(series)
+        self._assert_univariate(series)
         series = self.training_series
 
         # determine trend
@@ -332,7 +333,7 @@ class FFT(LocalForecastingModel):
 
         return self
 
-    def predict(self, n: int, num_samples: int = 1):
+    def predict(self, n: int, num_samples: int = 1, verbose: bool = False):
         super().predict(n, num_samples)
         trend_forecast = np.array(
             [self.trend_function(i + len(self.training_series)) for i in range(n)]