+++ darts/models/forecasting/forecasting_model.py
@@ -146,8 +146,6 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         self
             Fitted model.
         """
-        if not isinstance(self, FutureCovariatesLocalForecastingModel):
-            series._assert_univariate()
         raise_if_not(
             len(series) >= self.min_train_series_length,
             "Train series only contains {} elements but {} model requires at least {} entries".format(
@@ -243,8 +241,9 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         past_covariates: Optional[TimeSeries],
         future_covariates: Optional[TimeSeries],
         num_samples: int,
+        verbose: bool = False,
     ) -> TimeSeries:
-        return self.predict(n, num_samples=num_samples)
+        return self.predict(n, num_samples=num_samples, verbose=verbose)
 
     @property
     def min_train_series_length(self) -> int:
@@ -516,6 +515,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
                 past_covariates=past_covariates,
                 future_covariates=future_covariates,
                 num_samples=num_samples,
+                verbose=verbose,
             )
 
             if last_points_only:
@@ -561,7 +561,10 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         retrain: Union[bool, int, Callable[..., bool]] = True,
         overlap_end: bool = False,
         last_points_only: bool = False,
-        metric: Callable[[TimeSeries, TimeSeries], float] = metrics.mape,
+        metric: Union[
+            Callable[[TimeSeries, TimeSeries], float],
+            List[Callable[[TimeSeries, TimeSeries], float]],
+        ] = metrics.mape,
         reduction: Union[Callable[[np.ndarray], float], None] = np.mean,
         verbose: bool = False,
     ) -> Union[float, List[float]]:
@@ -637,9 +640,12 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         last_points_only
             Whether to use the whole historical forecasts or only the last point of each forecast to compute the error
         metric
-            A function that takes two ``TimeSeries`` instances as inputs and returns an error value.
+            A function or a list of function that takes two ``TimeSeries`` instances as inputs and returns an
+            error value.
         reduction
             A function used to combine the individual error scores obtained when `last_points_only` is set to False.
+            When providing several metric functions, the function will receive the argument `axis = 0` to obtain single
+            value for each metric function.
             If explicitly set to `None`, the method will return a list of the individual error scores instead.
             Set to ``np.mean`` by default.
         verbose
@@ -664,14 +670,26 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
             verbose=verbose,
         )
 
+        if not isinstance(metric, list):
+            metric = [metric]
+
         if last_points_only:
-            return metric(series, forecasts)
+            errors = [metric_f(series, forecasts) for metric_f in metric]
 
-        errors = [metric(series, forecast) for forecast in forecasts]
-        if reduction is None:
+        else:
+            # metric in columns, forecast in rows
+            errors = [
+                [metric_f(series, forecast) for metric_f in metric]
+                for forecast in forecasts
+            ]
+            if reduction is not None:
+                # one value per metric
+                errors = reduction(np.array(errors), axis=0)
+
+        if len(metric) > 1:
             return errors
-
-        return reduction(np.array(errors))
+        else:
+            return errors[0]
 
     @classmethod
     def gridsearch(
@@ -860,6 +878,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
                     metric=metric,
                     reduction=reduction,
                     last_points_only=last_points_only,
+                    verbose=verbose,
                 )
             else:  # split mode
                 model._fit_wrapper(series, past_covariates, future_covariates)
@@ -869,6 +888,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
                     past_covariates,
                     future_covariates,
                     num_samples=1,
+                    verbose=verbose,
                 )
                 error = metric(val_series, pred)
 
@@ -1207,6 +1227,12 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
 
         return model
 
+    def _assert_univariate(self, series: TimeSeries):
+        if not series.is_univariate:
+            raise_log(
+                ValueError("This model only supports univariate TimeSeries instances")
+            )
+
 
 class LocalForecastingModel(ForecastingModel, ABC):
     """The base class for "local" forecasting models, handling only single univariate time series.
@@ -1216,7 +1242,7 @@ class LocalForecastingModel(ForecastingModel, ABC):
     the entire target series supplied when calling :func:`fit()` at once. They can also predict in one go with
     :func:`predict()` for any number of predictions `n` after the end of the training series.
 
-    All implementations must implement the `_fit()` and `_predict()` methods.
+    All implementations must implement the `fit()` and `predict()` methods.
     """
 
     def __init__(self, add_encoders: Optional[dict] = None):
@@ -1234,6 +1260,11 @@ class LocalForecastingModel(ForecastingModel, ABC):
         Optional[List[int]],
     ]:
         return None, None, False, False, None, None
+
+    @abstractmethod
+    def fit(self, series: TimeSeries) -> "LocalForecastingModel":
+        super().fit(series)
+        series._assert_deterministic()
 
 
 class GlobalForecastingModel(ForecastingModel, ABC):
@@ -1319,6 +1350,7 @@ class GlobalForecastingModel(ForecastingModel, ABC):
         past_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
         future_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ) -> Union[TimeSeries, Sequence[TimeSeries]]:
         """Forecasts values for `n` time steps after the end of the series.
 
@@ -1385,6 +1417,7 @@ class GlobalForecastingModel(ForecastingModel, ABC):
         past_covariates: Optional[TimeSeries],
         future_covariates: Optional[TimeSeries],
         num_samples: int,
+        verbose: bool = False,
     ) -> TimeSeries:
         return self.predict(
             n,
@@ -1392,6 +1425,7 @@ class GlobalForecastingModel(ForecastingModel, ABC):
             past_covariates=past_covariates,
             future_covariates=future_covariates,
             num_samples=num_samples,
+            verbose=verbose,
         )
 
     def _fit_wrapper(
@@ -1564,6 +1598,8 @@ class FutureCovariatesLocalForecastingModel(LocalForecastingModel, ABC):
         n: int,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
+        **kwargs,
     ) -> TimeSeries:
         """Forecasts values for a certain number of time steps after the end of the series.
         DualCovariatesModels must implement the predict logic in this method.
@@ -1585,9 +1621,13 @@ class FutureCovariatesLocalForecastingModel(LocalForecastingModel, ABC):
         past_covariates: Optional[TimeSeries],
         future_covariates: Optional[TimeSeries],
         num_samples: int,
+        verbose: bool = False,
     ) -> TimeSeries:
         return self.predict(
-            n, future_covariates=future_covariates, num_samples=num_samples
+            n,
+            future_covariates=future_covariates,
+            num_samples=num_samples,
+            verbose=verbose,
         )
 
     @property
@@ -1767,6 +1807,7 @@ class TransferableFutureCovariatesLocalForecastingModel(
         historic_future_covariates: Optional[TimeSeries] = None,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ) -> TimeSeries:
         """Forecasts values for a certain number of time steps after the end of the series.
         TransferableFutureCovariatesLocalForecastingModel must implement the predict logic in this method.
@@ -1780,12 +1821,14 @@ class TransferableFutureCovariatesLocalForecastingModel(
         past_covariates: Optional[TimeSeries],
         future_covariates: Optional[TimeSeries],
         num_samples: int,
+        verbose: bool = False,
     ) -> TimeSeries:
         return self.predict(
             n=n,
             series=series,
             future_covariates=future_covariates,
             num_samples=num_samples,
+            verbose=verbose,
         )
 
     def _supports_non_retrainable_historical_forecasts(self) -> bool: