+++ darts/models/forecasting/sf_ets.py
@@ -78,7 +78,7 @@ class StatsForecastETS(FutureCovariatesLocalForecastingModel):
 
     def _fit(self, series: TimeSeries, future_covariates: Optional[TimeSeries] = None):
         super()._fit(series, future_covariates)
-        series._assert_univariate()
+        self._assert_univariate(series)
         series = self.training_series
         self.model.fit(
             series.values(copy=False).flatten(),
@@ -91,6 +91,7 @@ class StatsForecastETS(FutureCovariatesLocalForecastingModel):
         n: int,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ):
         super()._predict(n, future_covariates, num_samples)
         forecast_df = self.model.predict(