+++ darts/models/forecasting/varima.py
@@ -133,6 +133,7 @@ class VARIMA(TransferableFutureCovariatesLocalForecastingModel):
         historic_future_covariates: Optional[TimeSeries] = None,
         future_covariates: Optional[TimeSeries] = None,
         num_samples: int = 1,
+        verbose: bool = False,
     ) -> TimeSeries:
 
         if num_samples > 1 and self.trend: