+++ darts/tests/models/forecasting/test_local_forecasting_models.py
@@ -206,7 +206,7 @@ class LocalForecastingModelsTestCase(DartsBaseTestClass):
     def test_multivariate_input(self):
         es_model = ExponentialSmoothing()
         ts_passengers_enhanced = self.ts_passengers.add_datetime_attribute("month")
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(ValueError):
             es_model.fit(ts_passengers_enhanced)
         es_model.fit(ts_passengers_enhanced["#Passengers"])
         with self.assertRaises(KeyError):