+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Darts is still in an early development phase, and we cannot always guarantee bac
 ## [Unreleased](https://github.com/unit8co/darts/tree/master)
 - Fixed edge case in ShapExplainer for regression models where covariates series > target series [#1310](https://https://github.com/unit8co/darts/pull/1310) by [Rijk van der Meulen](https://github.com/rijkvandermeulen)
 - Removed `IPython` as a dependency. [#1331](https://github.com/unit8co/darts/pull/1331) by [Erik Hasse](https://github.com/erik-hasse)
+- New models: `DLinearModel` and `NLinearModel` as proposed in [this paper](https://arxiv.org/pdf/2205.13504.pdf). [#1139](https://github.com/unit8co/darts/pull/1139) by [Julien Herzen](https://github.com/hrzn) and [Greg DeVos](https://github.com/gdevos010).
 
 [Full Changelog](https://github.com/unit8co/darts/compare/0.22.0...master)